Add existing file

+++ src/TrashGTAGame/Block.java
@@ -16,11 +16,11 @@ public class Block extends GameObject {
     @Override
     public void render(Graphics g) {
         g.setColor(Color.BLACK);
-        g.fillRect(x, y, 32, 32);
+        g.fillRect((int)x, (int)y, 32, 32);
     }
 
     @Override
     public Rectangle getBounds() {
-        return new Rectangle(x, y, 32, 32);
+        return new Rectangle((int)x, (int)y, 32, 32);
     }
 }
